[KYUUBI #1505] Move ServiceNodeInfo to a separate file
